Create handleOk with necessary DB updates
Created handle for OK button with necessary SQL queries. This meant creating a new method to MovieDao to update the movie_actors joint table and making ID fetch method public so it can be used in ActorDialogViewController. At the moment, program doesn't update the Movie object itself, so the up-to-date list doesn't appear in the Main view. Added also few actors for testing.