Refactor actor auto-complete logic
Replaced direct database queries with in-memory filtering for actor auto-completion in initialize method. Now, ActorDao's readAll method is called in initializeController to fetch and store all actors in allActors list.
The auto-completion logic in initialize method now filters this list based on user input, limited to the top 10 matches.

This approach reduces database queries, improving performance and responsiveness. Although it might not be the best permanent solution with big databases.

Also with these changes removed the unnecessary findActorsByStartingNae from ActorDao.